+++ docs-website/sidebars.ts
@@ -69,6 +69,7 @@ const sidebars: SidebarsConfig = {
         'reference/type_expect',
         'reference/bitvec',
         'reference/lua_utils',
+        'reference/bundle_to_vlbc',
         'reference/str_bits_utils',
         'reference/symbol_helper',
         'reference/await_time',