+++ src/wave_vpi/xmake.lua
@@ -137,7 +137,7 @@ target("wave_vpi_wellen_impl", function()
     set_kind("phony")
     on_build(function()
         os.cd(path.join(curr_dir, "wellen_impl"))
-        os.exec("cargo build --release")
+        os.exec("cargo build -p wave_vpi_wellen_impl --release")
     end)
 
     after_build(function(target)